All unnecessary files for serial communication are removed.
Originally I was using Arduino uno for driving coils.
Unfortunately, Uno was not supported by micro-ros, and I was
going to reinvent a square wheel by writing a
serialization/deserialization protocol.

Fortunately, I found a Arduino Due in the lab that is supported
by micro-ros, and just finished Arduino side.
I chose not to write the history for two reason:
1- This is how the project was developed, let's keeps it.
2- This was my first project with Git, and I think I made some
mistakes, that may take some time to rewrite the history.
I do not have that time now, maybe I do it in the future.