Rotating in place independently is available now.
This change lets ut to command in-place rotation to robots through a
usual command in the current mode. Note that this is different than
lining up robots that can be executed automatically beforeeach section
of movements.

"feedforward_walk", "pivot_walking", and "pivot_walking_alt" methods in
ControlModel class of controller module are modified.